feat: 검색페이지 post list컴포넌트 수정[#30]
