+++ src/app/(tab)/search/page.tsx
@@ -3,7 +3,7 @@
 import SearchForm from "@/components/post/search-form";
 import { useFormState } from "react-dom";
 import { searchPosts } from "./action";
-import PostList from "@/components/post/post-list";
+import SearchedPostList from "@/components/post/search-post-list";
 
 export default function SearchPage() {
   const [state, action] = useFormState(searchPosts, null);
@@ -11,7 +11,7 @@ export default function SearchPage() {
   return (
     <div className="flex flex-col">
       <SearchForm onSearch={action} />
-      {state?.isSuccess ? <PostList initialPosts={state.data ?? []} /> : <p>{state?.message}</p>}
+      <SearchedPostList searchPosts={state?.data} />
     </div>
   );
 }
