+++ src/components/post/post-list.tsx
@@ -1,7 +1,7 @@
-import { CursorPagination, InitialPosts } from "@/service/postService";
+import { ListOfPost } from "@/lib/types";
 import ListPost from "./list-post";
 
-export default function PostList({ initialPosts }: { initialPosts: CursorPagination["items"] }) {
+export default function PostList({ initialPosts }: { initialPosts: ListOfPost[] }) {
   if (initialPosts.length === 0)
     return <div className="mx-auto py-10 text-xl text-underline">인증글이 존재하지 않습니다.</div>;
   return (
