+++ src/lib/types.ts
@@ -1,3 +1,5 @@
+import { AiBot, AiComment, Post, User } from "@prisma/client";
+
 export type ServerResponse<Data> = SuccessResponse<Data> | FailResponse;
 
 export type SuccessResponse<Data> = {
@@ -13,3 +15,13 @@ export type FailResponse = {
   error: Error;
   message: string;
 };
+
+type PostOfAiComment = AiComment & { aiBot: AiBot };
+export type ListOfPost = Post & { user: User } & { _count: { comments: number; likes: number } } & {
+  aiComments: PostOfAiComment[];
+};
+export interface CursorPagination {
+  items: ListOfPost[];
+  nextCursorId: number | null;
+  isLastPage: boolean;
+}
